% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytometryDiscrete}
\alias{cytometryDiscrete}
\title{The cytometry dataset}
\format{An \code{\link{data.frame}} with 11 variables and 5400 observations.}
\usage{
cytometryDiscrete
}
\description{
A discretized version of the dataset from \href{http://science.sciencemag.org/content/308/5721/523.long}{Sachs et al. (2005)}.
}
\keyword{datasets}

