% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsebn-plotting.R
\name{plotDAG}
\alias{plotDAG}
\title{Plot a DAG}
\usage{
plotDAG(x)
}
\arguments{
\item{x}{An \code{\link{edgeList}}, \code{\link{sparsebnFit}}, or \code{\link{sparsebnPath}} object.}

\item{...}{Additional arguments to \code{\link{plot}}.}
}
\description{
Using some sensible defaults for large graphs, plot a DAG object. Uses \code{\link[igraph]{igraph}}
package by default.
}
\details{
This method is not intended for customization. For more control over the output, use
\code{\link{plot}} and see \code{\link{setPlotPackage}} for plotting only and/or
\code{\link{setGraphPackage}} for even more control. These methods grants the user the full
feature set of the selected package.
}

