% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathfinder}
\alias{pathfinder}
\title{The pathfinder network}
\format{An \code{\link{edgeList}} object with 109 nodes and 195 edges.}
\usage{
pathfinder
}
\description{
The pathfinder network from the \href{http://www.bnlearn.com/bnrepository/#pathfinder}{Bayesian network repository}.
Pathfinder is an expert system developed in 1992 by Heckerman et. al
to assist with the diagnosis of lymph-node diseases.
}
\keyword{datasets}

